restructure DB

+++ dictionary/tables/IngredientType.sql
@@ -2,6 +2,6 @@ CREATE TABLE IF NOT EXISTS dictionary.ingredienttype
 (
     id              SMALLSERIAL  NOT NULL
         CONSTRAINT pk_ingredienttype PRIMARY KEY,
-    storageblock_id SMALLSERIAL  NOT NULL,
+    storageblock_id INTEGER  NOT NULL,
     name            VARCHAR(100) NOT NULL
 );
